Be more lenient about what makes a valid chunk key.
When a key is 9, 10, 13, or 14 bytes long it might be either a chunk key
or a plain key. If it contain only printable characters represent it in
chunk key form, even if the game would not read it as a chunk key.
This is better at future proofing the game.