Call digest() with raw=FALSE as xxhash64 did not have raw mode (#14)
But now / in the next version it does so we need the old behavior.

In digest 0.6.37 as on CRAN, raw is indifferent between TRUE or FALSE

```r
> library(digest)
> x <- "the quick brown fox jumped over the fence"
> digest::digest(x, algo = "xxhash64", serialize = FALSE, raw = TRUE)
[1] "0acd1be66972fa22"
> digest::digest(x, algo = "xxhash64", serialize = FALSE, raw = FALSE)
[1] "0acd1be66972fa22"
> packageVersion("digest")
[1] ‘0.6.37’
>
```

In digest as currently in its main branch and a CRAN candidate, it matters

```r
> library(digest)
> x <- "the quick brown fox jumped over the fence"
> digest::digest(x, algo = "xxhash64", serialize = FALSE, raw = TRUE)
[1] 0a cd 1b e6 69 72 fa 22
> digest::digest(x, algo = "xxhash64", serialize = FALSE, raw = FALSE)
[1] "0acd1be66972fa22"
> packageVersion("digest")
[1] ‘0.6.37.2’
>
```

With raw=FALSE we get the behaviour your currently rely on.